+++ source/core/src/main/com/csse3200/game/components/combat/CombatButtonDisplay.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
-import com.csse3200.game.areas.CombatArea;
+import com.csse3200.game.areas.combat.CombatArea;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.csse3200.game.entities.Entity;
 import com.csse3200.game.services.DialogueBoxService;
